+++ __tests__/flow.js
@@ -1,4 +1,4 @@
-/* @flow */
+// @flow
 
 export function something(arg: Something): string {
   return arg.blah