+++ react.js
@@ -1,4 +1,6 @@
 module.exports = {
+  parser: "babel-eslint",
+
   plugins: [
     "react",
   ],